Improve errors when an item no longer exists (#354)
If a user keeps a ref to an item, and that item disappears, calling a
method on that item will now raise a helpful StaleProxyError, instead of
a NoMethodError on nil
   
The proxy code was significantly refactored, and Thing will also inherit
the same stale proxy I also removed the $things registry lookup on every
access of a thing, using the same event subscriber mechanism as Item.
    
I also moved several methods from GenericItem to Item, so that Proxy can
detect that those methods should exist even for dummy/stale items.

Signed-off-by: Cody Cutrer <cody@cutrer.us>