Require irb ~> 1.14 to allow the version shipped with JRuby 10 (#439)
As #438 still proved to be difficult, this tackles it a different way.

JRuby 10 was just released, and with it, comes irb 1.14.3, which our
curent irb code is happy with. Since the console is only available in
openHAB 5, we could just reduce this version requirement to match
JRuby10, and ship OH5 with JRuby10. That would also solve the irb issue.

Signed-off-by: Jimmy Tanagra <jcode@tanagra.id.au>